+++ src/main/java/Selenide/BaseTestSelenide.java
@@ -1,10 +1,8 @@
 package Selenide;
 
-import com.codeborne.selenide.Condition;
-import com.codeborne.selenide.SelenideElement;
+
 import lombok.extern.log4j.Log4j2;
 import properties.PropertyReader;
-
 import static com.codeborne.selenide.Selenide.open;
 import static com.codeborne.selenide.Selenide.page;
 import static com.codeborne.selenide.WebDriverRunner.driver;
@@ -20,15 +18,5 @@ public class BaseTestSelenide {
         return driver().hasWebDriverStarted() ? page(pageObjectClassClass) : open(link, pageObjectClassClass);
     }
 
-    protected void isDisplayed(SelenideElement ... elements) {
-        for (SelenideElement element : elements) {
-            element.shouldBe(Condition.exist);
-            log.info(element + " exists");
-        }
-    }
-
-    protected void getText(SelenideElement element) {
-        element.text();
-    }
 
 }
