+++ src/main/java/Selenide/Moodpanda/AboutUsPage.java
@@ -1,10 +1,11 @@
 package Selenide.Moodpanda;
 
+import Selenide.BasePage;
 import Selenide.BaseTestSelenide;
 import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.support.FindBy;
 
-public class AboutUsPage extends BaseTestSelenide {
+public class AboutUsPage extends BasePage {
 
     @FindBy(xpath = "//*[h2]")
     SelenideElement title;
