+++ src/main/java/Selenide/Moodpanda/FaqPage.java
@@ -1,10 +1,11 @@
 package Selenide.Moodpanda;
 
+import Selenide.BasePage;
 import Selenide.BaseTestSelenide;
 import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.support.FindBy;
 
-public class FaqPage extends BaseTestSelenide {
+public class FaqPage extends BasePage {
 
     @FindBy(tagName = "h2")
     SelenideElement title;
