+++ src/main/java/Selenide/Moodpanda/LoginPage.java
@@ -1,12 +1,13 @@
 package Selenide.Moodpanda;
 
+import Selenide.BasePage;
 import Selenide.BaseTestSelenide;
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
 import lombok.extern.log4j.Log4j2;
 import org.openqa.selenium.support.FindBy;
 @Log4j2
-public class LoginPage extends BaseTestSelenide {
+public class LoginPage extends BasePage {
 
     @FindBy(css = "#UpdateMoodWelcome .modal-header h4")
     SelenideElement loginPageTitle;
