+++ src/main/java/Selenide/Moodpanda/OverviewPage.java
@@ -1,5 +1,6 @@
 package Selenide.Moodpanda;
 
+import Selenide.BasePage;
 import Selenide.BaseTestSelenide;
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.ElementsCollection;
@@ -7,7 +8,7 @@ import com.codeborne.selenide.SelenideElement;
 
 import static com.codeborne.selenide.Selenide.$$;
 
-public class OverviewPage extends BaseTestSelenide {
+public class OverviewPage extends BasePage {
 
     ElementsCollection aboutElements = $$(".about-content h3");
